+++ data/stations.csv
@@ -1,3 +1,3 @@
 station_name, location, longitude, latitude, altitude, start_bin_height, end_bin_height, n_bins, dt, gdas1_folder, gdastxt_folder, lidar_src_folder,molecular_dataset,lidar_dataset,db_file,aeronet_folder,aeronet_name,generation_folder,gen_signal_dataset,gen_lidar_dataset,gen_aerosol_dataset,gen_bg_dataset,gen_density_dataset
 haifa, Haifa,35.02,32.8,230,78.75,22485.66016,3000, 50 * 1E-9, H:\data_haifa\DATA FROM TROPOS\GDAS\haifa, H:\data_haifa\DATA FROM TROPOS\GDAS\haifa_preproc, H:\data_haifa\DATA FROM TROPOS\data\level1a\PollyXT_TROPOS,D:\data_haifa\DATA FROM TROPOS\molecular_dataset,D:\data_haifa\DATA FROM TROPOS\lidar_dataset,H:\data_haifa\DATA FROM TROPOS\data\level1a\PollyXT_TROPOS\pollyxt_tropos_calibration.db,D:\data_haifa\AERONET,Technion_Haifa_IL,D:\data_haifa\GENERATION\params_dataset,D:\data_haifa\GENERATION\signal_dataset,D:\data_haifa\GENERATION\lidar_dataset,D:\data_haifa\GENERATION\aerosol_dataset,D:\data_haifa\GENERATION\bg_dataset,D:\data_haifa\GENERATION\density_dataset
-haifa_shubi, Haifa,35.02,32.8,230,78.75,22485.66016,3000, 50 * 1E-9, data_example/gdas, data_example/gdas_txt, DATA FROM TROPOS,data_example/molecular_dataset,D:\data_haifa\DATA FROM TROPOS\lidar_dataset,data_example/pollyxt_tropos_calibration.db,D:\data_haifa\AERONET,Technion_Haifa_IL,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/bg_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/signal_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/lidar_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/aerosol_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/bg_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/density_dataset
+haifa_shubi, Haifa,35.02,32.8,230,78.75,22485.66016,3000, 50 * 1E-9, data_example/gdas, data_example/gdas_txt, /home/shubi/PycharmProjects/learning_lidar/data/DATA FROM TROPOS,data_example/molecular_dataset,D:\data_haifa\DATA FROM TROPOS\lidar_dataset,data_example/pollyxt_tropos_calibration.db,D:\data_haifa\AERONET,Technion_Haifa_IL,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/bg_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/signal_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/lidar_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/aerosol_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/bg_dataset,/home/shubi/PycharmProjects/learning_lidar/data/GENERATION/density_dataset
