+++ learning_lidar/generation/generate_LC_pattern.py
@@ -11,6 +11,7 @@ import xarray as xr
 import torch.utils.data
 
 # Local modules
+import learning_lidar.preprocessing.preprocessing_utils as prep_utils
 import learning_lidar.utils.global_settings as gs
 import learning_lidar.preprocessing.preprocessing as prep
 from learning_lidar.utils.proc_utils import Bezier
@@ -178,8 +179,8 @@ def main(station_name, start_date, end_date):
     n_pts = p_slice.Time.size
     t0 = p_slice.Time[0].values
     t1 = p_slice.Time[-1].values
-    dt0 = prep.dt64_2_datetime(t0)
-    dt1 = prep.dt64_2_datetime(t1)
+    dt0 = prep_utils.dt64_2_datetime(t0)
+    dt1 = prep_utils.dt64_2_datetime(t1)
     difft = (dt1 - dt0)
 
     n_total = difft.days * station.total_time_bins + difft.seconds / station.freq
