+++ learning_lidar/generation/generation_utils.py
@@ -3,6 +3,8 @@ import learning_lidar.preprocessing.preprocessing as prep
 from datetime import datetime, timedelta
 import os
 import calendar
+
+import learning_lidar.preprocessing.preprocessing_utils as prep_utils
 from learning_lidar.utils.global_settings import TIMEFORMAT
 from tqdm import tqdm
 import pandas as pd
@@ -52,7 +54,7 @@ def save_generated_dataset(station, dataset, data_source='lidar', save_mode='bot
     :param profiles: The name of profile desired to be saved separately.
     :return: ncpaths - the paths of the saved dataset/s . None - for failure.
     """
-    date_datetime = prep.get_daily_ds_date(dataset)
+    date_datetime = prep_utils.get_daily_ds_date(dataset)
     if data_source == 'lidar':
         base_folder = station.gen_lidar_dataset
     elif data_source == 'signal':
@@ -63,9 +65,9 @@ def save_generated_dataset(station, dataset, data_source='lidar', save_mode='bot
         base_folder = station.gen_density_dataset
     elif data_source == 'bg':
         base_folder = station.gen_bg_dataset
-    month_folder = prep.get_month_folder_name(base_folder, date_datetime)
+    month_folder = prep_utils.get_month_folder_name(base_folder, date_datetime)
 
-    prep.get_daily_ds_date(dataset)
+    prep_utils.get_daily_ds_date(dataset)
     '''save the dataset to separated netcdf files: per profile per wavelength'''
     ncpaths = []
 
@@ -122,7 +124,7 @@ def get_month_gen_params_path(station, day_date, type='density_params'):
     nc_name = f"generated_{type}_{station.location}_{month_start_day.strftime('%Y-%m-%d')}_" \
               f"{month_end_day.strftime('%Y-%m-%d')}.nc"
 
-    folder_name = prep.get_month_folder_name(station.generation_folder, day_date)
+    folder_name = prep_utils.get_month_folder_name(station.generation_folder, day_date)
 
     gen_source_path = os.path.join(folder_name, nc_name)
     return gen_source_path
@@ -209,7 +211,7 @@ def plot_daily_profile(profile_ds, height_slice=None, figsize=(16, 6)):
 def plot_hourly_profile(profile_ds, height_slice=None, figsize=(10, 6), times=None):
     # TODO : move to vis_utils.py
     # TODO: add scientific ticks on color-bar
-    day_date = prep.dt64_2_datetime(profile_ds.Time[0].values)
+    day_date = prep_utils.dt64_2_datetime(profile_ds.Time[0].values)
     str_date = day_date.strftime("%Y-%m-%d")
     if times == None:
         times = [day_date + timedelta(hours=8),
